api/WebXPloitApi/requirements.txt
aiofiles==24.1.0
Cython==3.0.11
fastapi==0.115.0
pydantic==2.9.2
tortoise-orm==0.17.2
uvicorn==0.31.0
cython
